feat: directory and files renaming
